Keyboard and mouse movement event listeners are working
They bind to document (for keyboard), and canvas (for mouse movement),
passing the event details to the wasm handlers, which are then firing
correctly.

Weirdly, the keyboard event listener doesn't seem to want to bind
to the canvas.  But it's working when bound to the entire document,
which is good enough for now, and might be better.  Not sure why
it's not binding the canvas though.